🔨 refactor: mv client-compat to client
