+++ src/ducks/settings/GroupsSettings.jsx
@@ -3,7 +3,7 @@ import { withRouter } from 'react-router'
 import { translate, Button, Icon } from 'cozy-ui/react'
 import Table from 'components/Table'
 import { ACCOUNT_DOCTYPE, GROUP_DOCTYPE } from 'doctypes'
-import { queryConnect } from 'utils/client-compat'
+import { queryConnect } from 'utils/client'
 import Loading from 'components/Loading'
 import plus from 'assets/icons/16/plus.svg'
 import styles from './GroupsSettings.styl'
