+++ src/utils/client.js
@@ -1,37 +1,86 @@
 /* global __TARGET__ */
 
-import { initClient } from 'ducks/authentication/lib/client'
-import { CozyClient } from 'old-cozy-client'
+import { initMobileStackClient } from 'ducks/authentication/lib/client'
 
-export const getClientMobile = persistedState => {
-  const hasPersistedMobileStore = persistedState && persistedState.mobile
-  return initClient(hasPersistedMobileStore ? persistedState.mobile.url : '')
-}
+import CozyStackClient from 'cozy-stack-client'
+import CozyClient, { StackLink, withMutations, Query } from 'cozy-client'
+import PouchLink from 'cozy-pouch-link'
+import { schema } from 'doctypes'
+import { flowRight as compose } from 'lodash'
 
-export const getClientBrowser = () => {
+import React from 'react'
+
+export const links = {}
+
+export const getClient = async () => {
   const root = document.querySelector('[role=application]')
   const data = root.dataset
-  return new CozyClient({
-    cozyURL: `${window.location.protocol}//${data.cozyDomain}`,
-    token: data.cozyToken
+
+  const stackClient = __TARGET__ === 'mobile'
+    ? await initMobileStackClient()
+    : new CozyStackClient({
+      uri: `${window.location.protocol}//${data.cozyDomain}`,
+      token: data.cozyToken
+    })
+
+  if (__TARGET__ === 'mobile') {
+    links.pouch = new PouchLink({
+      doctypes: ['io.cozy.bank.operations'],
+      // doctypes: [ACCOUNT_DOCTYPE, GROUP_DOCTYPE, TRANSACTION_DOCTYPE]
+      initialSync: true,
+      client: stackClient
+    })
+  }
+
+  links.stack = new StackLink({ client: stackClient })
+
+  const client = new CozyClient({
+    link: [
+      __TARGET__ === 'mobile' && links.pouch,
+      links.stack
+    ].filter(Boolean),
+    schema: schema,
+    // TODO: we need to pass the stack client to the client so that
+    // OAuth register() call can be forwarded to the right stack client...
+    // Definitely not good...
+    client: stackClient
   })
+
+  return client
 }
 
-const memoize = fn => {
-  let res
-  return (...args) => {
-    if (typeof res === 'undefined') {
-      res = fn(...args)
+export const withQuery = (dest, queryOpts) => Component => (props, context) => {
+  queryOpts = typeof queryOpts === 'function' ? queryOpts(props) : queryOpts
+  if (queryOpts.doc) {
+    return <Component {...{ [dest]: queryOpts.doc, ...props }} />
+  } else {
+    if (!context.client) {
+      console.warn('Context', context)
+      throw new Error(
+        'Query should be used with client in context (use CozyProvider to set context)'
+      )
     }
-    return res
+    return (
+      <Query {...queryOpts}>
+        {result => {
+          console.log('query result', dest, result)
+          return <Component {...{ [dest]: result, ...props }} />
+        }}
+      </Query>
+    )
   }
 }
 
-export const getClient = memoize(persistedState => {
-  return __TARGET__ === 'mobile'
-    ? getClientMobile(persistedState)
-    : getClientBrowser()
-})
+export const withCrud = withMutations(client => ({
+  saveDocument: document => client.save(document),
+  destroyDocument: document => client.destroy(document)
+}))
 
-export const isCollectionLoading = col =>
-  col && (col.fetchStatus === 'loading' || col.fetchStatus === 'pending')
+export const queryConnect = querySpecs => Component => {
+  const enhancers = Object.keys(querySpecs).map(dest =>
+    withQuery(dest, querySpecs[dest])
+  )
+  enhancers.push(withCrud)
+  return compose.apply(null, enhancers)(Component)
+}
+
